+++ src/domain/providers/npm/npmConfig.ts
@@ -3,10 +3,7 @@ import type { HttpOptions } from '#domain/clients';
 import type { IFrozenOptions } from '#domain/configuration';
 import type { IProviderConfig } from '#domain/providers';
 import { NpmFeatures } from '#domain/providers/npm';
-import { nameOf } from '#domain/utils';
 import { throwUndefinedOrNull } from '@esm-test/guards';
-
-const def = nameOf<NpmConfig>();
 
 export class NpmConfig implements IProviderConfig {
 
@@ -15,9 +12,9 @@ export class NpmConfig implements IProviderConfig {
     readonly caching: CachingOptions,
     readonly http: HttpOptions
   ) {
-    throwUndefinedOrNull(def.config, config);
-    throwUndefinedOrNull(def.caching, caching);
-    throwUndefinedOrNull(def.http, http);
+    throwUndefinedOrNull('config', config);
+    throwUndefinedOrNull('caching', caching);
+    throwUndefinedOrNull('http', http);
   }
 
   readonly fileLanguage = ['yaml', 'json', 'jsonc'];