+++ src/domain/providers/npm/npmSuggestionProvider.ts
@@ -14,13 +14,10 @@ import {
   type JsonPackageTypeHandler,
   type JsonParserOptions,
   type PackageNameDescriptor,
+  type PackagePathDescriptor,
   type PackageVersionDescriptor,
-  type YamlParserOptions,
   PackageDescriptorType,
-  createVersionDescFromJsonNode,
-  createVersionDescFromYamlNode,
-  parsePackagesJson,
-  parsePackagesYaml
+  parsePackagesJson
 } from '#domain/parsers';
 import type { ISuggestionProvider } from '#domain/providers';
 import {
@@ -33,6 +30,7 @@ import {
   convertNpmErrorToResponse,
   createNpmRegistryClientData,
   createNpmSuggestionFromErrorCode,
+  createNpmVersionDescFromJsonNode,
   customDescriptorHandler,
   npmReplaceVersion,
   resolveDotFilePath
@@ -44,12 +42,8 @@ import { resolve } from 'node:path';
 import npa from 'npm-package-arg';
 
 const jsonComplexTypeHandlers: KeyDictionary<JsonPackageTypeHandler> = {
-  [PackageDescriptorType.version]: createVersionDescFromJsonNode
+  [PackageDescriptorType.version]: createNpmVersionDescFromJsonNode
 };
-
-const yamlComplexTypeHandlers = {
-  [PackageDescriptorType.version]: createVersionDescFromYamlNode
-}
 
 export class NpmSuggestionProvider implements ISuggestionProvider {
 
@@ -70,11 +64,72 @@ export class NpmSuggestionProvider implements ISuggestionProvider {
   parseDependencies(
     packagePath: string,
     packageText: string,
-    dependencyProperties = this.config.dependencyProperties
+    dependencyProps = this.config.dependencyProperties
   ): Array<PackageDependency> {
-    return packagePath.endsWith('pnpm-workspace.yaml')
-      ? parsePnpmWorkspace(packagePath, packageText, dependencyProperties)
-      : parsePackageJson(packagePath, packageText, dependencyProperties)
+    const options: JsonParserOptions = {
+      includePropNames: dependencyProps,
+      complexTypeHandlers: jsonComplexTypeHandlers,
+      customDescriptorHandler
+    };
+
+    const parsedPackages = parsePackagesJson(packageText, options);
+    const packageDependencies = [];
+    for (const descriptors of parsedPackages) {
+      const nameDesc = descriptors.getType<PackageNameDescriptor>(
+        PackageDescriptorType.name
+      );
+
+      // handle any pnpm override dependency selectors in the name
+      let name = nameDesc.name;
+      const atIndex = name.indexOf('@');
+      if (atIndex > 0) {
+        name = name.slice(0, atIndex);
+      }
+
+      // map the version descriptor to a package dependency
+      if (descriptors.hasType(PackageDescriptorType.version)) {
+        const versionDesc = descriptors.getType<PackageVersionDescriptor>(
+          PackageDescriptorType.version
+        );
+
+        if (['catalog:', 'workspace:'].some(x => versionDesc.version.startsWith(x))) {
+          continue;
+        }
+
+        packageDependencies.push(
+          new PackageDependency(
+            createPackageResource(
+              name,
+              versionDesc.version,
+              packagePath
+            ),
+            descriptors
+          )
+        );
+        continue;
+      }
+
+      // map the path descriptor to a package dependency
+      if (descriptors.hasType(PackageDescriptorType.path)) {
+        const pathType = descriptors.getType<PackagePathDescriptor>(
+          PackageDescriptorType.path
+        );
+        packageDependencies.push(
+          new PackageDependency(
+            createPackageResource(
+              nameDesc.name,
+              pathType.path,
+              packagePath
+            ),
+            descriptors
+          )
+        );
+        continue;
+      }
+
+    }
+
+    return packageDependencies;
   }
 
   async preFetchSuggestions(projectPath: string, packagePath: string): Promise<NpmClientData> {
@@ -175,94 +230,4 @@ export class NpmSuggestionProvider implements ISuggestionProvider {
 
   }
 
-}
-
-function parsePackageJson(packagePath: string, packageText: string, dependencyProps: string[]) {
-  const options: JsonParserOptions = {
-    includePropNames: dependencyProps,
-    complexTypeHandlers: jsonComplexTypeHandlers,
-    customDescriptorHandler
-  };
-
-  const parsedPackages = parsePackagesJson(packageText, options);
-  const packageDependencies = [];
-  for (const descriptors of parsedPackages) {
-    const nameDesc = descriptors.getType<PackageNameDescriptor>(
-      PackageDescriptorType.name
-    );
-
-    // handle any pnpm override dependency selectors in the name
-    let name = nameDesc.name;
-    const atIndex = name.indexOf('@');
-    if (atIndex > 0) {
-      name = name.slice(0, atIndex);
-    }
-
-    // map the version descriptor to a package dependency
-    if (descriptors.hasType(PackageDescriptorType.version)) {
-      const versionDesc = descriptors.getType<PackageVersionDescriptor>(
-        PackageDescriptorType.version
-      );
-
-      if (['catalog:', 'workspace:'].some(x => versionDesc.version.startsWith(x))) {
-        continue;
-      }
-
-      packageDependencies.push(
-        new PackageDependency(
-          createPackageResource(
-            name,
-            versionDesc.version,
-            packagePath
-          ),
-          descriptors
-        )
-      );
-
-      continue;
-    }
-
-  }
-
-  return packageDependencies;
-}
-
-function parsePnpmWorkspace(packagePath: string, packageText: string, dependencyProps: string[]) {
-  const options: YamlParserOptions = {
-    includePropNames: dependencyProps,
-    complexTypeHandlers: yamlComplexTypeHandlers
-  };
-
-  const parsedPackages = parsePackagesYaml(packageText, options);
-
-  const packageDependencies = [];
-
-  for (const descriptors of parsedPackages) {
-    const nameDesc = descriptors.getType<PackageNameDescriptor>(
-      PackageDescriptorType.name
-    );
-
-    // map the version descriptor to a package dependency
-    if (descriptors.hasType(PackageDescriptorType.version)) {
-      const versionType = descriptors.getType<PackageVersionDescriptor>(
-        PackageDescriptorType.version
-      );
-
-      packageDependencies.push(
-        new PackageDependency(
-          createPackageResource(
-            nameDesc.name,
-            versionType.version,
-            packagePath
-          ),
-          descriptors
-        )
-      );
-
-      //   continue;
-    }
-
-  } // end map loop
-
-  return packageDependencies;
 }