+++ test/unit/domain/providers/all.tests.ts
@@ -7,5 +7,6 @@ export * as DubProvider from './dub/index.tests';
 export * as GoProvider from './golang/index.tests';
 export * as MavenProvider from './maven/index.tests';
 export * as NpmProvider from './npm/index.tests';
+export * as PnpmProvider from './pnpm/index.tests';
 export * as PubProvider from './pub/index.tests';
 export * as PypiProvider from './pypi/index.tests';