+++ test/unit/domain/providers/npm/npmSuggestionProvider.fixtures.ts
@@ -93,42 +93,6 @@ q/I2+0j6dAkOGcK/68z7qQXByeGri3n28a1Kn6o=
           ),
         ],
       }
-    },
-    yaml: {
-      test: `
-        catalog:
-          react: ^16.14.0
-        catalogs:
-          react17:
-            react: ^17.0.2
-            react-dom: ^17.0.2
-      `,
-      expected: [
-        new PackageDependency(
-          createPackageResource('react', '^16.14.0', 'test/path/pnpm-workspace.yaml'),
-          new PackageDescriptor([
-            createPackageNameDesc('react', createTextRange(28)),
-            createPackageVersionDesc('^16.14.0', createTextRange(35, 43)),
-            createPackageParentDescType('catalog'),
-          ])
-        ),
-        new PackageDependency(
-          createPackageResource('react', '^17.0.2', 'test/path/pnpm-workspace.yaml'),
-          new PackageDescriptor([
-            createPackageNameDesc('react', createTextRange(93)),
-            createPackageVersionDesc('^17.0.2', createTextRange(100, 107)),
-            createPackageParentDescType('catalogs.*.*'),
-          ])
-        ),
-        new PackageDependency(
-          createPackageResource('react-dom', '^17.0.2', 'test/path/pnpm-workspace.yaml'),
-          new PackageDescriptor([
-            createPackageNameDesc('react-dom', createTextRange(120)),
-            createPackageVersionDesc('^17.0.2', createTextRange(131, 138)),
-            createPackageParentDescType('catalogs.*.*'),
-          ])
-        ),
-      ]
     }
   }
 }